disable the *extremely* inefficient `Symbol` manipulation
Since `Fcf.Text` performs `Uncons` by converting a `Symbol` into a
`List`, which involves comparing it to `Symbol`s in a table to get each
character, it inflicts an absolutely insane penalty on compile times.
Doing that with 20 `Symbols` takes more than five seconds.

Instead, use the old variant again, which delays this conversion to the
consumers of the typelevel field names, which simply compare
`a == "_" <> b`.